chore: extract useKeyboardHeight hook into another file
